django/requirements.txt
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
dj-static==0.0.6
Django==3.1.4
django-anymail==8.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.1
iniconfig==1.1.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.6.0
openpyxl==3.0.5
packaging==20.7
pluggy==0.13.1
psycopg2==2.9
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
zipp==3.4.0
pillow==8.1.0
